refactor(brainstorm): split monolith into modular sub-package
Replace single 746-line brainstorm.py with a dedicated sub-package
containing separate modules for orthodox/unorthodox generation,
analysis, citation regeneration, research handling, and core logic.
Improves maintainability and testability while preserving CLI behavior.
