+++ litassist/commands/brainstorm/__init__.py
@@ -1,0 +1,10 @@
+"""
+Brainstorm command module for generating legal strategies.
+
+This module provides comprehensive legal strategy generation including
+orthodox, unorthodox, and most-likely-to-succeed analysis.
+"""
+
+from .core import brainstorm
+
+__all__ = ["brainstorm"]
